Cleaned up setup.sh and docker image from pcibus renderer
