+++ deployments/nvml-mock/scripts/setup.sh
@@ -372,7 +372,6 @@ if [ -x /usr/local/bin/check-fabric ]; then
   cp /usr/local/bin/check-fabric "$DRIVER_ROOT/usr/bin/check-fabric"
 fi
 cp -a /usr/local/lib/libibmock*.so* "$DRIVER_ROOT/usr/local/lib/" 2>/dev/null || true
-cp -a /usr/local/lib/libpcimocksys.so* "$DRIVER_ROOT/usr/local/lib/" 2>/dev/null || true
 
 # Runtime overrides (written by nvml-mock-ctl) are ephemeral: wipe them on
 # every pod start so a restart of this DaemonSet resets simulated GPU state
@@ -392,98 +391,9 @@ if [ -f /etc/nvml-mock/topology/topology.yaml ]; then
   echo "Staged ComputeDomain topology overlay at $HOST/topology/topology.yaml"
 fi
 
-# 7. Label node (requires RBAC: get+patch on nodes, for gpu.present only)
-#
-#    nvidia.com/gpu.present is written directly: it is an NVIDIA-namespaced key
-#    no NFD source derives.
-#
-#    feature.node.kubernetes.io/pci-10de.present is NOT written directly. We
-#    drop a feature file and let NFD create the label, so that key has exactly
-#    one writer and "NFD works" stays distinguishable from "we set it
-#    ourselves" (#505).
-#
-#    Why not NFD's PCI source? As deployed it cannot see our devices.
-#    (Upstream facts as of NFD v0.19.0, the version pinned in go.mod; the
-#    worker actually deployed comes from GPU Operator and may differ.)
-#    source/pci/utils.go resolves hostpath.SysfsDir.Path("bus/pci/devices"),
-#    and pkg/utils/hostpath builds SysfsDir from a private `pathPrefix` set
-#    only via linker -X. Upstream's container build passes
-#    HOSTMOUNT_PREFIX=/host- (Makefile), so the shipped worker reads
-#    /host-sys/bus/pci/devices — the real host /sys, hostPath-mounted by the
-#    worker DaemonSet. Our tree is at /var/lib/nvml-mock/sys and is reachable
-#    only through libpcimocksys.so, whose injection into the worker is inert
-#    twice over:
-#      a) the shim rewrites only paths starting "/sys/" (k_prefixes[] in
-#         pkg/system/mockpcisysfs/c/shim.c), so "/host-sys/..." never matches.
-#      b) nfd-worker is built -extldflags=-static, so it has no PT_INTERP and
-#         ignores LD_PRELOAD outright.
-#    The key itself is right: GPU Operator configures NFD with
-#    deviceLabelFields:[vendor] and whitelists class 0302, which is exactly the
-#    "pci-<vendor>.present" form, and our rendered devices carry all five
-#    attributes NFD treats as mandatory. Only visibility is missing.
-#
-#    NFD's *local* source has no such limit: source/local/local.go reads
-#    featureFilesDir, a plain literal (not host-prefixed), and the worker
-#    mounts that host directory at the same path. Each line parses as
-#    key[=value]. That is the supported route, and it needs no node RBAC. The
-#    GFD mock already writes to that same directory (tests/e2e/gfd-mock.yaml:52).
-#
-#    MOCK_NFD_PCI_LABEL (chart value `nodeLabels.pciVendorPresent`, default
-#    true/on) gates only the feature file. Turn it off on a cluster where
-#    something else already supplies the key. This step converges either way:
-#    with the gate on it writes the file, with the gate off it removes any
-#    file an earlier run left behind. cleanup.sh reads the same variable and
-#    unwinds only the write this step makes — with the gate off it has nothing
-#    to remove, because the removal already happened here.
-PCI_LABEL_MODE=$(printf '%s' "${MOCK_NFD_PCI_LABEL:-on}" | tr '[:upper:]' '[:lower:]')
-case "$PCI_LABEL_MODE" in
-  off | on) ;;
-  *)
-    echo "ERROR: MOCK_NFD_PCI_LABEL='$MOCK_NFD_PCI_LABEL' is invalid; expected off or on" >&2
-    exit 1
-    ;;
-esac
+# 7. Label node with nvidia.com/gpu.present (requires RBAC: get+patch on nodes).
 if command -v kubectl >/dev/null 2>&1; then
   kubectl label node "$NODE_NAME" nvidia.com/gpu.present=true --overwrite || true
-fi
-
-# NFD local source: drop a feature file instead of writing the label
-# ourselves. source/local/local.go reads every file in featureFilesDir and
-# parses each line as key[=value], so this one line becomes
-# feature.node.kubernetes.io/pci-10de.present=true — created by NFD, not by
-# us. Requires no node RBAC, unlike the kubectl label call it replaces.
-# With no NFD on the cluster the file is inert and the label simply does not
-# exist, which is the honest state (#505).
-NFD_FEATURES_DIR=/host/etc/kubernetes/node-feature-discovery/features.d
-NFD_FEATURE_FILE="$NFD_FEATURES_DIR/nvml-mock.features"
-if [ "$PCI_LABEL_MODE" = "on" ]; then
-  # Tolerant on purpose — a failure here must not be fatal, as with the `mknod`
-  # calls in step 3, the `cp` calls in step 4b and the `kubectl label` above.
-  # Those three discard the failure with `|| true`; this one reports it on
-  # stderr. It replaced a `kubectl label` call here that carried `|| true`
-  # (cleanup.sh carried the matching one for the removal), and under
-  # `set -e` (line 9) a bare failure here would abort the entrypoint before
-  # the IB and PCI renders in steps 9 and 10, crash-looping the whole mock for
-  # an optional, gated feature. When the write fails no label appears, which is
-  # the honest state (#505), and nothing downstream is corrupted — unlike the
-  # IB and PCI renders in steps 9 and 10, which are deliberately fatal because
-  # a partial tree silently misleads its consumers. A failing `if` CONDITION
-  # does not trip `set -e`, so this form warns and continues rather than
-  # swallowing the error the way `|| true` would.
-  if mkdir -p "$NFD_FEATURES_DIR" && echo "pci-10de.present=true" > "$NFD_FEATURE_FILE"; then
-    echo "Wrote $NFD_FEATURE_FILE (NFD derives feature.node.kubernetes.io/pci-10de.present)"
-  else
-    echo "WARNING: could not write $NFD_FEATURE_FILE — feature.node.kubernetes.io/pci-10de.present will not be created; the mock is otherwise unaffected" >&2
-  fi
-else
-  # Mirror of the tolerant write above: the removal is the same optional,
-  # gated feature on its other arm, so an EROFS or unwritable directory must
-  # not abort the entrypoint before steps 9 and 10 either.
-  if rm -f "$NFD_FEATURE_FILE"; then
-    echo "Skipping NFD pci-10de feature file (nodeLabels.pciVendorPresent=false)"
-  else
-    echo "WARNING: could not remove $NFD_FEATURE_FILE — a stale feature file may keep feature.node.kubernetes.io/pci-10de.present alive; the mock is otherwise unaffected" >&2
-  fi
 fi
 
 # 8b. Deliberately NOT written here: /run/nvidia/validations/toolkit-ready.
@@ -574,24 +484,7 @@ if [ "$MOCK_IB_MODE" != "off" ] && [ -x /usr/local/bin/mock-ib ]; then
   fi
 fi
 
-# 10. Render fake PCI sysfs tree (consumed by topology-aware DRA / device
-#     plugins that resolve PCIe root complex via a readlink on
-#     /sys/bus/pci/devices/<bdf>). The renderer parses the profile's
-#     `pcie_topology:` block; profiles without one get a flat default
-#     covering every device under a single root complex (`pci0000:00`,
-#     NUMA 0). Failures are fatal under `set -e` for the same reason as
-#     the IB block above — a topology typo otherwise yields silently
-#     malformed sysfs that downstream `dra.k8s.io/pcieRoot` attributes
-#     would inherit.
-PCI_ROOT="$HOST"
-mkdir -p "$PCI_ROOT"
-if [ -x /usr/local/bin/render-pci-sysfs ]; then
-  /usr/local/bin/render-pci-sysfs \
-    --config /etc/nvml-mock/config.yaml \
-    --output "$PCI_ROOT"
-fi
-
-# 11. Fabric Manager: on NVSwitch platforms (HGX H100 / GB200 / GB300) the
+# 10. Fabric Manager: on NVSwitch platforms (HGX H100 / GB200 / GB300) the
 #     real nvidia-fabricmanager registers the GPUs with the NVSwitch fabric
 #     before they are usable. When MOCK_FABRICMANAGER is enabled we start the
 #     fake daemon, which writes a node-local readiness marker under
